+++ testing/test_column_selector_2d.py
@@ -1,11 +1,12 @@
 import numpy as np
 import pytest
 
-from neuraxle.steps.column_transformer import ColumnSelector2D, ColumnsSelectorND
+from neuraxle.steps.column_transformer import ColumnSelector2D, ColumnsSelectorND, NumpyColumnSelector2D
 
 
-def test_column_selector_2d_should_select_range():
-    step = ColumnSelector2D(range(0, 10))
+@pytest.mark.parametrize('column_selector_2d_class', [ColumnSelector2D, NumpyColumnSelector2D])
+def test_column_selector_2d_should_select_range(column_selector_2d_class):
+    step = column_selector_2d_class(range(0, 10))
     data_inputs, expected_outputs = _create_data_source((20, 20))
 
     outputs = step.transform(data_inputs)
@@ -13,8 +14,9 @@ def test_column_selector_2d_should_select_range():
     assert np.array_equal(outputs, data_inputs[..., :10])
 
 
-def test_column_selector_2d_should_select_int():
-    step = ColumnSelector2D(10)
+@pytest.mark.parametrize('column_selector_2d_class', [ColumnSelector2D, NumpyColumnSelector2D])
+def test_column_selector_2d_should_select_int(column_selector_2d_class):
+    step = column_selector_2d_class(10)
     data_inputs, expected_outputs = _create_data_source((20, 20))
 
     outputs = step.transform(data_inputs)
@@ -23,8 +25,9 @@ def test_column_selector_2d_should_select_int():
     assert np.array_equal(outputs, expected_data_inputs)
 
 
-def test_column_selector_2d_should_select_slice():
-    step = ColumnSelector2D(slice(0, 10, 1))
+@pytest.mark.parametrize('column_selector_2d_class', [ColumnSelector2D, NumpyColumnSelector2D])
+def test_column_selector_2d_should_select_slice(column_selector_2d_class):
+    step = column_selector_2d_class(slice(0, 10, 1))
     data_inputs, expected_outputs = _create_data_source((20, 20))
 
     outputs = step.transform(data_inputs)
@@ -32,18 +35,19 @@ def test_column_selector_2d_should_select_slice():
     assert np.array_equal(outputs, data_inputs[..., :10])
 
 
-def test_column_selector_2d_should_select_list_of_indexes():
-    step = ColumnSelector2D([0, 1, 2])
+@pytest.mark.parametrize('column_selector_2d_class', [ColumnSelector2D, NumpyColumnSelector2D])
+def test_column_selector_2d_should_select_list_of_indexes(column_selector_2d_class):
+    step = column_selector_2d_class([0, 1, 2])
     data_inputs, expected_outputs = _create_data_source((20, 20))
 
     outputs = step.transform(data_inputs)
 
     assert np.array_equal(outputs, data_inputs[..., :3])
-    pass
 
 
-def test_column_selector_2d_should_throw_exception_on_unsupported_type():
-    step = ColumnSelector2D('unsupported')
+@pytest.mark.parametrize('column_selector_2d_class', [ColumnSelector2D, NumpyColumnSelector2D])
+def test_column_selector_2d_should_throw_exception_on_unsupported_type(column_selector_2d_class):
+    step = column_selector_2d_class('unsupported')
     data_inputs, expected_outputs = _create_data_source((20, 20))
 
     with pytest.raises(ValueError):
